Optimise pipeline: M1 config tuning, M2 mask/beamform improvements, M3 int8 whisper, M4 dedup, README rewrite
- M1: renumber scripts (04-06), remove dead code, clean up doa_estimate
  duplicate saves, fix dual-map loading in track_and_cluster
- M2: add early dedup via DoA posterior correlation, competitive soft
  masking with adaptive sigma, subband masks, manifest system for
  candidate tracking, stale file cleanup
- M3: switch Whisper CPU compute to int8 (prevents OOM on 14 GB RAM),
  add stale analysis cleanup on re-run
- M4: add transcript-based dedup (Jaccard + bigram), proximity filter
  for provisionals, stale output cleanup
- config: tuned enhancement params (mask_exponent=3, null_strength=2.5,
  subband_smooth_bins=16, pitch threshold=165 Hz)
- README: rewrite with installation, member descriptions, run instructions
- .gitignore: cover sweep PNGs, intermediate JSON, sweep calibration
- Remove unused audio_utils.py, trim plotting/json_schema/constants

+++ src/member3_analysis/02_asr_whisper.py
@@ -99,7 +99,7 @@ def transcribe(
             has_cuda = False
 
         device = "cuda" if has_cuda else "cpu"
-        compute_type = "float16" if has_cuda else "float32"
+        compute_type = "float16" if has_cuda else "int8"
         logger.info("Loading Whisper model '%s' (device=%s, compute=%s)",
                      model_size, device, compute_type)
         _whisper_model_cache["model"] = WhisperModel(
